@@ -1,6 +1,6 @@
 module github.com/nspcc-dev/neofs-contract
 
-go 1.21
+go 1.22
 
 require (
 	github.com/google/uuid v1.6.0
