@@ -66,7 +66,7 @@ func BenchmarkDivideFundsEvenly(b *testing.B) {
 			b.ReportAllocs()
 			b.ResetTimer()
 
-			for i := 0; i < b.N; i++ {
+			for range b.N {
 				divideFundsEvenly(tc.a, tc.n, func(ind int, amount uint64) {})
 			}
 		})
